+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   `ColorVariable`
 * Gracefully handle when a library specified via `-r`/`--require` is not found
   and return a semantic exit code `EX_UNAVAILABLE` (`69`)
+* Improve message of `DeclarationOrder` to specify the line number of the
+  out-of-order item and the line number of where it was expected to be
 
 ## 0.35.0